Add options and display modules
