+++ display.rb
@@ -1,0 +1,12 @@
+module Display
+  def display_menu
+    puts "\nPlease choose an option by entering a number:"
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Add a person'
+    puts '4 - Add a book'
+    puts '5 - Add a rental'
+    puts '6 - List all rentals for a given person id'
+    puts '7 - Exit'
+  end
+end
