+++ input.rb
@@ -1,0 +1,9 @@
+class Input
+  def input
+    gets.chomp
+  end
+
+  def input_int
+    gets.chomp.to_i
+  end
+end
